requirements.txt
opencv-python~=4.7.0.72
pillow~=9.5.0
numpy~=1.24.3
urllib3==1.26.7
moviepy==1.0.3
pydub~=0.25.1
python-docx
docx2txt
google-cloud-texttospeech
google~=3.0.0
pyttsx3
django~=4.2.2
stanfordnlp~=0.2.0
python-pptx
Pillow
requests~=2.31.0
transformers
stanfordcorenlp
deep